上傳圖片修改完成

+++ cakeshoping/pages/admin/products/[id].js
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import ProductItem from '../../../components/ProductItem';
+import Updating from '../../../components/Updating';
 import {
-  getProducts,
   getProduct,
   getPhoto,
   deletePhoto,
@@ -11,9 +11,12 @@ import {
 } from '../../../pages/api/webAPI';
 import { server } from '../../../config';
 import { useRouter } from 'next/router';
+import { useSelector } from 'react-redux';
+import { selectAdminUser } from '../../../features/adminUserSlice';
+import NotFound from '../../../components/NotFound';
 
 export default function Product({ productData, photosData }) {
-  const router = useRouter();
+  const adminUser = useSelector(selectAdminUser);
   const [productName, setProductName] = useState(productData.productName);
   const [type, setType] = useState(productData.type);
   const [price, setPrice] = useState(productData.price);
@@ -21,19 +24,36 @@ export default function Product({ productData, photosData }) {
   const [isShow, setIsShow] = useState(productData.isShow ? true : false);
   const [storage, setStorage] = useState(productData.storage);
   const [sell, setSell] = useState(productData.sell);
-  const [id, setID] = useState(productData.id);
   const [isDeleted, setIsDeleted] = useState(productData.isDeleted);
   const [photos, setPhotos] = useState(photosData);
   const [photoSrc, setPhotoSrc] = useState([]);
+  const [isUpdating, setIsUpdating] = useState(false);
+  const [errorMessage, setErrorMessage] = useState(false);
 
-  const handleUploadFile = async (e) => {
+  const router = useRouter();
+  const { id } = router.query;
+  if (!id) return <></>;
+
+  const handleUploadFile = (e) => {
     let fromData = new FormData();
     for (let i = 0; i < e.target.files.length; i++) {
       fromData.append('avatar', e.target.files[i]);
-      fromData.append('productid', id);
     }
-    await addNewPhoto(fromData).catch((error) => console.log(error));
-    console.log(fromData);
+    fromData.append('productId', id);
+    setIsUpdating(true);
+    addNewPhoto(fromData).then((res) => {
+      if (!res) {
+        setIsUpdating(false);
+        console.log(res);
+        return setErrorMessage(true);
+      }
+      if (res.statusText === 'OK') {
+        setIsUpdating(false);
+        getPhoto(id).then((res) => {
+          setPhotos(res.data.result);
+        });
+      }
+    });
   };
 
   const handleUpdateProduct = async (data) => {
@@ -42,8 +62,11 @@ export default function Product({ productData, photosData }) {
   };
 
   const handleDeletePhoto = (id) => {
-    setPhotos(photos.filter((photo) => photo.id !== id));
-    deletePhoto(id);
+    let deleteMessage = window.confirm('確定刪除圖片？');
+    if (deleteMessage) {
+      setPhotos(photos.filter((photo) => photo.id !== id));
+      deletePhoto(id);
+    }
   };
   const handleDeleteProduct = (id) => {
     deleteProduct(id);
@@ -55,31 +78,37 @@ export default function Product({ productData, photosData }) {
 
   return (
     <>
-      <ProductItem
-        title={`編輯商品 ID:`}
-        photos={photos}
-        productName={productName}
-        type={type}
-        price={price}
-        articlel={articlel}
-        isShow={isShow}
-        storage={storage}
-        sell={sell}
-        id={id}
-        setProductName={setProductName}
-        setType={setType}
-        setPrice={setPrice}
-        setArticlel={setArticlel}
-        setStorage={setStorage}
-        setSell={setSell}
-        isDeleted={isDeleted}
-        handleUpdateProduct={handleUpdateProduct}
-        handleDeletePhoto={handleDeletePhoto}
-        handleDeleteProduct={handleDeleteProduct}
-        handleIsShowClick={handleIsShowClick}
-        photoSrc={photoSrc}
-        handleUploadFile={handleUploadFile}
-      />
+      {isUpdating && <Updating />}
+      {errorMessage && <div>發生錯誤請再試一次</div>}
+      {adminUser.role === 'admin' ? (
+        <ProductItem
+          title={`編輯商品 ID:`}
+          photos={photos}
+          productName={productName}
+          type={type}
+          price={price}
+          articlel={articlel}
+          isShow={isShow}
+          storage={storage}
+          sell={sell}
+          id={id}
+          setProductName={setProductName}
+          setType={setType}
+          setPrice={setPrice}
+          setArticlel={setArticlel}
+          setStorage={setStorage}
+          setSell={setSell}
+          isDeleted={isDeleted}
+          handleUpdateProduct={handleUpdateProduct}
+          handleDeletePhoto={handleDeletePhoto}
+          handleDeleteProduct={handleDeleteProduct}
+          handleIsShowClick={handleIsShowClick}
+          photoSrc={photoSrc}
+          handleUploadFile={handleUploadFile}
+        />
+      ) : (
+        <NotFound />
+      )}
     </>
   );
 }
@@ -93,6 +122,7 @@ export const getStaticProps = async (content) => {
       productData: productData.result[0],
       photosData: photosData.result,
     },
+    revalidate: 60,
   };
 };
